change types

+++ src/views/NotesView.vue
@@ -5,8 +5,8 @@
       <div class="actions">
         <h2>{{ pageObject?.name }}</h2>
         <div>
-          <div class="tooltip tooltip tooltip-left" data-tip="Adicionar nova nota">
-            <button class="" @click="addNewNoteModal">
+          <div class="tooltip tooltip-left" data-tip="Adicionar nova nota">
+            <button @click="addNewNoteModal">
               <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5"
                 stroke="currentColor" class="size-6">
                 <path stroke-linecap="round" stroke-linejoin="round" d="M12 4.5v15m7.5-7.5h-15" />
@@ -14,9 +14,8 @@
             </button>
           </div>
 
-
-          <div class="tooltip tooltip tooltip-left" data-tip="Remover todas as notas">
-            <button class="" @click="confirmDeleteAllSubItems">
+          <div class="tooltip tooltip-left" data-tip="Remover todas as notas">
+            <button @click="confirmDeleteAllSubItems">
               <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5"
                 stroke="currentColor" class="size-6">
                 <path stroke-linecap="round" stroke-linejoin="round"
@@ -30,7 +29,6 @@
       <div class="content" @click="contentClicked">
         <div v-for="subitem in subitems" :key="subitem.id" :data-id="subitem.id" :data-slug="subitem.slug" class="note">
           <div class="actions">
-
             <svg class="remove-note size-4" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24"
               stroke-width="1.5" stroke="currentColor">
               <path stroke-linecap="round" stroke-linejoin="round"
@@ -42,8 +40,6 @@
               <path stroke-linecap="round" stroke-linejoin="round"
                 d="m16.862 4.487 1.687-1.688a1.875 1.875 0 1 1 2.652 2.652L10.582 16.07a4.5 4.5 0 0 1-1.897 1.13L6 18l.8-2.685a4.5 4.5 0 0 1 1.13-1.897l8.932-8.931Zm0 0L19.5 7.125M18 14v4.75A2.25 2.25 0 0 1 15.75 21H5.25A2.25 2.25 0 0 1 3 18.75V8.25A2.25 2.25 0 0 1 5.25 6H10" />
             </svg>
-
-
           </div>
           <h3>{{ subitem.title }}</h3>
           <p class="content">{{ subitem.content }}</p>
@@ -54,11 +50,6 @@
       </div>
     </div>
   </div>
-
-
-
-
-
 
   <dialog id="confirm_delete_all_subitems" class="modal">
     <div class="modal-box">
@@ -72,14 +63,11 @@
       </div>
       <div class="modal-action">
         <form method="dialog">
-          <!-- if there is a button in form, it will close the modal -->
           <button class="btn" @click="deleteAllSubItems">Remover</button>
         </form>
       </div>
     </div>
   </dialog>
-
-
 
   <dialog id="add_new_card" class="modal">
     <div class="modal-box">
@@ -92,22 +80,17 @@
           <span class="label-text">Título da nota</span>
         </div>
         <input type="text" name="title" class="input input-bordered w-full max-w-xs" />
-
         <div class="label">
           <span class="label-text">Tags</span>
         </div>
         <input type="text" name="tags" class="input input-bordered w-full max-w-xs" />
-
-
         <div class="label">
           <span class="label-text">Texto da nota</span>
         </div>
         <textarea rows="5" name="content" class="textarea textarea-bordered"></textarea>
-
       </div>
       <div class="modal-action">
         <form method="dialog">
-          <!-- if there is a button in form, it will close the modal -->
           <button class="btn" @click="addNewNote">Adicionar</button>
         </form>
       </div>
@@ -115,19 +98,36 @@
   </dialog>
 </template>
 
-
 <script setup lang="ts">
-import { ref, onMounted, type DialogHTMLAttributes } from 'vue'
+import { ref, onMounted } from 'vue'
 import { useRouter } from 'vue-router'
 import useIDB from '@/utils/useIDB.ts'
 import useUtils from '@/utils/useUtils'
 
+type Subitem = {
+  id: number
+  title: string
+  slug: string
+  content: string
+  date: string
+  tags: string[]
+}
+
+type Page = {
+  id: number
+  name: string
+  slug: string
+  subitems: Subitem[]
+}
+
+let add_new_card: HTMLDialogElement
+let confirm_delete_all_subitems: HTMLDialogElement
+
 const utils = useUtils()
 const router = useRouter()
 const idbUtils = useIDB()
-const pageObject = ref(null)
-const subitems = ref([])
-
+const pageObject = ref<Page | null>(null)
+const subitems = ref<Subitem[]>([])
 
 function addNewNoteModal() {
   clearModal()
@@ -143,62 +143,63 @@ function clearModal() {
   content.value = ''
 }
 
-function setModalContent(subitem: any) {
+function setModalContent(subitem: Subitem) {
   const title = document.querySelector('input[name="title"]') as HTMLInputElement
   const tags = document.querySelector('input[name="tags"]') as HTMLInputElement
   const content = document.querySelector('textarea[name="content"]') as HTMLTextAreaElement
   const button = document.querySelector('.modal-action button') as HTMLButtonElement
-  const parent = button.closest('.modal-box')
+  const parent = button.closest('.modal-box') as HTMLDivElement
   button.textContent = 'Editar'
   button.classList.add('edit')
   title.value = subitem.title
   tags.value = subitem.tags.join(',')
-  parent.dataset.id = subitem.id
+  parent.dataset.id = String(subitem.id)
   content.value = subitem.content
 }
 
 async function addNewNote(event: Event) {
   const target = event.target as HTMLElement
-  const parent = target.closest('.modal-box')
+  const parent = target.closest('.modal-box') as HTMLDivElement
   if (!parent) return
   const title = parent.querySelector('input[name="title"]') as HTMLInputElement
   const tags = parent.querySelector('input[name="tags"]') as HTMLInputElement
   const content = parent.querySelector('textarea[name="content"]') as HTMLTextAreaElement
   const button = parent.querySelector('.modal-action button') as HTMLButtonElement
   if (!title.value || !content.value) return
-  const date = new Date();
-  const month = String(date.getMonth() + 1).padStart(2, '0');
-  const day = date.getDate();
-  const year = date.getFullYear();
-  const formattedDate = `${month}/${day}/${year}`;
+  const date = new Date()
+  const month = String(date.getMonth() + 1).padStart(2, '0')
+  const day = date.getDate()
+  const year = date.getFullYear()
+  const formattedDate = `${month}/${day}/${year}`
   const slug = utils.slugify(title.value)
   const tagsArray = tags.value.split(',')
 
-  const subitem = {
+  const newSubitem = {
     title: title.value,
-    slug: slug,
+    slug,
     content: content.value,
     date: formattedDate,
-    tags: tagsArray
-  }
+    tags: tagsArray,
+    id: 0,
+  } as Subitem
 
   if (button.classList.contains('edit')) {
     const id = Number(parent.dataset.id)
-    const updateSubItem = await idbUtils.updateSubItem(1, id, { ...subitem, id })
-    subitems.value = subitems.value.map((subitem) => {
-      if (subitem.id === updateSubItem.id) {
+    newSubitem.id = id
+    const updateSubItem = await idbUtils.updateSubItem(1, id, newSubitem)
+    subitems.value = subitems.value.map((item) => {
+      if (item.id === updateSubItem.id) {
         return updateSubItem
       }
-      return subitem
+      return item
     })
     button.classList.remove('edit')
     add_new_card.close()
     return
   }
 
-  const updateSubItem = await idbUtils.addSubItem(1, subitem)
-  subitems.value.push(updateSubItem)
-
+  const addedSubItem = await idbUtils.addSubItem(1, newSubitem)
+  subitems.value.push(addedSubItem)
 }
 
 function confirmDeleteAllSubItems() {
@@ -206,96 +207,84 @@ function confirmDeleteAllSubItems() {
 }
 
 async function deleteAllSubItems() {
+  if (!pageObject.value) return
   await idbUtils.deleteAllSubItems(pageObject.value.id)
   subitems.value = []
 }
 
 async function contentClicked(event: MouseEvent) {
   const target = event.target as HTMLElement
-  const note = target.closest('.note')
+  const note = target.closest('.note') as HTMLDivElement
   const removeNote = target.closest('.remove-note')
   const editNote = target.closest('.edit-note')
-  const id = Number(note?.dataset.id)
   if (!note) return
+  const id = Number(note.dataset.id)
   if (removeNote) {
-    const subitem = subitems.value.find((subitem) => subitem.id === id)
-    if (!subitem) return
-
-    await idbUtils.deleteSubItem(pageObject.value.id, subitem.id)
-    subitems.value = subitems.value.filter((subitem) => subitem.id !== id)
+    const item = subitems.value.find((el) => el.id === id)
+    if (!item) return
+    if (!pageObject.value) return
+    await idbUtils.deleteSubItem(pageObject.value.id, item.id)
+    subitems.value = subitems.value.filter((sub) => sub.id !== id)
     return
   }
   if (editNote) {
-    const subitem = subitems.value.find((subitem) => subitem.id === id)
-    if (!subitem) return
-    setModalContent(subitem)
+    const item = subitems.value.find((el) => el.id === id)
+    if (!item) return
+    setModalContent(item)
     add_new_card.showModal()
   }
-
-  // router.push(`/note/${slug}`)
 }
 
 onMounted(async () => {
-  const pageSlug = router.currentRoute.value.params.slug
+  add_new_card = document.getElementById('add_new_card') as HTMLDialogElement
+  confirm_delete_all_subitems = document.getElementById('confirm_delete_all_subitems') as HTMLDialogElement
+  const pageSlug = router.currentRoute.value.params.slug as string
   const pages = await idbUtils.getPages()
-  const page = pages.find((page) => page.slug === pageSlug)
-  if (!page) return
-  pageObject.value = page
-  subitems.value = await idbUtils.getSubItems(page.id)
-
+  const foundPage = pages.find((p: Page) => p.slug === pageSlug)
+  if (!foundPage) return
+  pageObject.value = foundPage
+  subitems.value = await idbUtils.getSubItems(foundPage.id)
 })
-
-
-
 </script>
 
 <style scoped>
 .inside-page {
   width: 100%;
   height: 100%;
-  /* padding: 1rem; */
 }
 
-div .header {
+.header {
   height: 60px;
   width: 100%;
   font-size: 1.3rem;
   display: flex;
   justify-content: flex-start;
   align-items: center;
-  /* padding-left: 1rem; */
   border-bottom: 1px solid #303030;
-
 }
 
-div.body {
+.body {
   height: calc(100% - 120px);
   padding: 0 1rem;
 }
 
-div .body>.actions {
+.body>.actions {
   display: flex;
   justify-content: space-between;
   align-items: center;
   padding: 1rem;
-  /* height: 60px; */
   border-bottom: 1px solid #303030;
-
 }
 
-
-div .body>.actions>div {
+.body>.actions>div {
   width: 50%;
   display: flex;
   justify-content: flex-end;
   gap: 1rem;
 }
 
-
-
-div .body>.content {
+.body>.content {
   padding: 1rem;
-  height: fit-content;
   display: grid;
   width: 100%;
   grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
@@ -326,8 +315,6 @@ div .body>.content {
   grid-template-rows: 30px calc(250px - 4rem) 20px;
   gap: 1rem;
   position: relative;
-
-
 }
 
 .note .actions {
@@ -336,13 +323,11 @@ div .body>.content {
   bottom: 10px;
   display: none;
   display: flex;
-
 }
 
 .note:hover .actions {
   display: flex;
 }
-
 
 .note .actions svg {
   cursor: pointer;
@@ -357,13 +342,11 @@ div .body>.content {
   color: #ff0000;
 }
 
-
 .note .actions svg.edit-note:hover {
   color: aqua;
 }
 
 .note .content {
-
   overflow-x: hidden;
   overflow-y: auto;
   white-space: break-spaces;
@@ -395,7 +378,6 @@ div .body>.content {
   height: fit-content;
 }
 
-
 .modal-content {
   height: 100%;
   padding: 1rem 0 1rem 0;
@@ -411,7 +393,6 @@ div .body>.content {
   max-width: unset;
   scrollbar-width: thin;
   resize: none;
-
 }
 
 .modal-action {
